+++ paper/paper.bib
@@ -94,7 +94,8 @@ Nachdem die groβe Bedeutung der Bildkraft für die Verfestigung von dispersions
 @inproceedings{Halpin1969,
  author = {Halpin, John C.},
  title = {Effects of Environmental Factors on Composite Materials},
- year = {1969}
+ year = {1969},
+ doi = {10.21236/ad0692481}
 }
 
 @article{Boehlke2001,
