+++ README.md
@@ -29,7 +29,7 @@ Currently, HomoPy offers two types of homogenization procedures:
 The Halpin-Tsai method is an empirical approach to homogenize two isotropic materials (cf. [[Halpin1969]](#Halpin1969)). Our approach is modified with the Shear-Lag model after Cox (cf. [[Cox1952]](#Cox1952)), which is also used in [[Fu2002]](#Fu2002) and [[Lauke2019]](#Lauke2019). Being a scalar homogenization scheme, it allows to calculate the effective stiffness in the plane which is orthogonal to the isotropic plane within transverse isotropic materials, as it is the case for unidirectional reinforced polymers. The effective stiffness, or Young's modulus, is then a function of the angle to the reinforcing direction. A fiber distrubtion within the plane is recognized by volume averaging of imaginary plies of individual orientations in analogy to the laminate theory.
 
 ## Mori-Tanaka
-The Mori-Tanaka scheme goes back to Mori and Tanaka (cf. [[Mori1973]](#Mori1973)) and is a mean-field homogenization scheme based on Eshelby's solution (cf. [[Eshelby1957]](#Eshelby1957)). The implementation so far only allows spheroidal inclusions, which in fact is an ellispoid with identical minor axes or ellipsoid of revolution, respectively. Our algorithm allows to homogenize materials with different types of fibers, each possibily having an individual fiber distrubtion. Being a tensorial homogenization scheme, the fiber orientation tensor is directly included in the calculation and the result is an effective stiffness tensor. The authors would like to emphasize that for multi-inclusion problems or non-isotropic inclusions, the effective stiffness tensor may violate thermodynamic requirements, such as a symmetric stiffness tensor. Further readings of this attribute are given in [[Qiu1990]](#Qiu1990) and [[Weng1990]](#Weng1990). To compensate this, HomoPy offers an algorithm introduced in [[Segura2023]](#Segura2023), which always results in symmetric effective stiffnesses.
+The Mori-Tanaka scheme goes back to Mori and Tanaka (cf. [[Mori1973]](#Mori1973)) and is a mean-field homogenization scheme based on Eshelby's solution (cf. [[Eshelby1957]](#Eshelby1957)). The implementation so far allows UD (needle), spheroidal and circular inclusions. Our algorithm allows to homogenize materials with different types of fibers/inclusions, each possibily having an individual orientation distrubtion. Being a tensorial homogenization scheme, the fiber orientation tensor is directly included in the calculation and the result is an effective stiffness tensor. The authors would like to emphasize that the classic formulation after [[Benveniste1987]](#Benveniste1987) results in an effective stiffness tensor which violates thermodynamic requirements, i.e. which does not contain the major symmetry, for when multi-inclusion materials or non-isotropic inclusions are used, respectively. Further readings on this attribute are given in [[Qiu1990]](#Qiu1990) and [[Weng1990]](#Weng1990). To compensate this, HomoPy offers an algorithm introduced in [[Segura2023]](#Segura2023), which always results in symmetric effective stiffnesses.
 
 ### Documentation
 
@@ -63,6 +63,8 @@ Further topic related methods:
 <a id="Lauke2019">[Lauke2019]</a> S.-Y. Fu, B. Lauke, Y.-W. Mai, *Science and engineering of short fibre-reinforced polymer composites*, Woodhead Publishing (2019). \
[…]
