+++ README.md
@@ -63,7 +63,7 @@ Further topic related methods:
 <a id="Lauke2019">[Lauke2019]</a> S.-Y. Fu, B. Lauke, Y.-W. Mai, *Science and engineering of short fibre-reinforced polymer composites*, Woodhead Publishing (2019). \
 <a id="Mori1973">[Mori1973]</a> T. Mori, K. Tanaka, *Average stress in matrix and average elastic energy of materials with misfitting inclusions*, Acta Metallurgica 21 (5) (1973), 571-574. \
 <a id="Eshelby1957">[Eshelby1957]</a> J.-D. Eshelby, *The determination of the elastic field of an ellipsoidal inclusion, and related problems*, Proceedings of the Royal Society of London A 241 (1957), 376–396. \
-<a id="Benveniste1987">[Benveniste1987]</a> Y. Benveniste, *A new approach to the application of Mori-Tanaka's theory in composite materials*, Mechanics of Materials (1989), pp. 147-157, https://doi.org/10.1016/0167-6636(87)90005-6 
+<a id="Benveniste1987">[Benveniste1987]</a> Y. Benveniste, *A new approach to the application of Mori-Tanaka's theory in composite materials*, Mechanics of Materials (1989), pp. 147-157, https://doi.org/10.1016/0167-6636(87)90005-6 \
 <a id="Qiu1990">[Qui1990]</a> Y. P. Qiu, G. J. Weng, *On the application of mori-tanaka’s theory involving transversely isotropic spheroidal inclusions*, International Journal of Engineering Science 28 (11) (1990) 1121-1137. doi:10.1016/00207225(90)90112-V. \
 <a id="Weng1990">[Weng1990]</a> G. J. Weng, *The theoretical connection between mori-tanaka’s theory and the hashin-shtrikman-walpole bounds*, International Journal of Engineering Science 28 (11) (1990) 1111–1120. doi:10.1016/00207225(90)90111-U \
 <a id="Segura2023">[Segura2023]</a> N. J. Segura, B. L.A. Pichler and C. Hellmich, *Concentration tensors preserving elastic symmetry of multiphase composites*, Mechanics of Materials 178 (2023), https://doi.org/10.1016/j.mechmat.2023.104555
