### 2023-03-22 #5199 @Kevin-Mattheus-Moerman
@nicoguaro, @likask, @lizarett, @nickelnine37, @frederickgosselin, @PolymerGuy, @fmatuschke
 would you be interested in reviewing this submission  for JOSS? It concerns a "Python package for homogenization procedures in fiber reinforced polymers".

[Paper ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05199/joss.05199/10.21105.joss.05199.pdf)
[Software ](https://github.com/Extraweich/homopy)
