### 2023-03-23 #5295 @editorialbot
[REVIEW]: HomoPy: A Python package for homogenization procedures in fiber reinforced polymers
**Submitting author:** @Extraweich (<a href="http://orcid.org/0000-0002-4713-8096">Nicolas Christ</a>)
**Repository:** https://github.com/Extraweich/homopy
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
