+++ paper/paper.md
@@ -28,7 +28,7 @@ affiliations:
     index: 2
 -   name: University of Waterloo, Waterloo, ON, Canada
     index: 3
-date: 20 February 2023
+date: 23 March 2023
 bibliography: paper.bib
 
 ---
@@ -47,7 +47,7 @@ A general challenge in using FRP in engineering is that predictive simulations r
 
 ![Schematic of implemented homogenization methods, where $\mathbb{C}_i$ is the stiffness tensor of component $i$, $\bar{\mathbb{C}}$ is the effective stiffness tensor, $\phi_i$ is the orientation angle of fiber $i$ and $a_i$ its aspect ratio. Illustration in reference to @Fu1998b. \label{fig:MTHT}](images/Schematic.png)
 
-`HomoPy` was developed to contain two commonly used homogenization methods, namely the Mori-Tanka (cf. @Mori1973) for 3D stiffness predictions and a shear-lag modified Halpin-Tsai method (cf. @Cox1952 and @Halpin1969) for planar predictions, i.e. laminate predictions, which is based on the formulations in @Fu2019. The goal of `HomoPy` is to provide an open-source implementation of these methods with a particular focus on hybrid FRP modeling. Other modules offering similar functionalities are available, e.g. `fiberpy` and `mechmean` (cf. @mechmean), but to the author's knowledge none of them provide the capability to model hybrid FRPs consisting of different fiber materials and/or geometries.\
+`HomoPy` was developed to contain two commonly used homogenization methods, that is, the Mori-Tanka (cf. @Mori1973) for 3D stiffness predictions and a shear-lag modified Halpin-Tsai method (cf. @Cox1952 and @Halpin1969) for planar predictions, i.e. laminate predictions, which is based on the formulations in @Fu2019. The goal of `HomoPy` is to provide an open-source implementation of these methods with a particular focus on hybrid FRP modeling. Other modules offering similar functionalities are available, e.g. `fiberpy` and `mechmean` (cf. @mechmean), but to the author's knowledge none of them provide the capability to model hybrid FRPs consisting of different fiber materials and/or geometries.\
 A key element in HomoPy is the implementation of the graphical representation of the effective directional stiffnesses according to @Boehlke2001. Comparing different material systems or FRP tape layup orientations is intuitively easier with a graphical representation than comparing up to 21 stiffness components each. The package `mechkit` (cf. @Bauer2022b) also offers this visualization procedure, but is further capable of visualizing the generalized bulk modulus. Other visualization tools can be found in the package `continuum-mechanics` (cf. @continuum_mechanics), though the visualization is limited to second rank tensors. To the authors' knowledge, only `HomoPy` offers a visualization tool for elastic tensors in laminates and 2D stiffness tensors in general, respectively.
 
 To this point, `HomoPy` is limited to calculating the effective elastic properties with the two methods mentioned above. Possible extensions for the future include thermal expansion properties and other homogenization methods.
