### 2023-03-24 #5295 @Extraweich
Hi everybody and thank you for your volunteering. I hope that we together can have a good review experience :)
Would it be possible to set the version to 1.0.11? There were some minor changes to the previous version. Concerning the contribution of @JulianKarlBauer, we agreed on not having him as a co-author, but mention his contribution in the acknowledgement section.
