### 2023-03-24 #5199 @JulianKarlBauer
> Julian

Thank you @Kevin-Mattheus-Moerman  for raising this question. @Extraweich  kindly contacted me beforehand and again after your remark. I only did a little bit of code clean-up and took the opportunity to compare some minor features which homopy shares with mechkit and mechmean. These comparisons actually made their way to the tests, which I am very happy about. However, I did not contribute to the actual new content within homopy nor the paper itself. Therefore, I think that co-authorship is not appropriate. Thank you again for raising the question :)
And thanks @Extraweich  for making this great contribution and giving the opportunity for comparisons, this is very valuable.
