### 2023-04-01 #5295 @nicoguaro
> @nicoguaro, you can find them in the [docs](https://homopy.readthedocs.io/en/latest/index.html). Do they need to be in the README in the repository, too?

I could not find the instruction in the docs.
