### 2023-04-02 #5295 @Extraweich
> > @nicoguaro, you can find them in the [docs](https://homopy.readthedocs.io/en/latest/index.html). Do they need to be in the README in the repository, too?
>
> I could not find the instruction in the docs.

Am I right in assuming that you have found them by now? :)
