### 2023-04-21 #5295 @lizarett
The paper/documentation appears to be well-written in general. The main comments currently are for the authors to describe in more details (in the documentation) the actual limitations of the implemented mathematical formulations, including geometrical shapes of initial material samples, number of materials allowed, relevant characteristics, etc. Some papers on the Mori-Tanaka method indicate a limited range of validity in terms of several parameters. The documentation of HomoPy uses "alternative forms", and it would be helpful to clarify, what kind of limitations are currently in place for the user. This may reiterate the published literature (papers, not code libraries), but, imo, it is worth it.
