+++ paper/paper.bib
@@ -291,3 +291,16 @@ abstract = {The tensile properties of hybrid, injection molded, short glass fibr
  url = {https://github.com/nicoguaro/continuum_mechanics},
  doi = {10.5281/zenodo.4029448}
 }
+
+@software{fiberoripy,
+  author       = {Nils Meyer and
+                  kraussco and
+                  JulianKarlBauer},
+  title        = {meyer-nils/fiberoripy: v1.1.0},
+  month        = nov,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {v1.1.0},
+  doi          = {10.5281/zenodo.7305587},
+  url          = {https://doi.org/10.5281/zenodo.7305587}
+}
