+++ paper/paper.md
@@ -48,12 +48,14 @@ A general challenge in using FRP in engineering is that predictive simulations r
 ![Schematic of implemented homogenization methods, where $\mathbb{C}_i$ is the stiffness tensor of component $i$, $\bar{\mathbb{C}}$ is the effective stiffness tensor, $\phi_i$ is the orientation angle of fiber $i$ and $a_i$ its aspect ratio. Illustration in reference to @Fu1998b. \label{fig:MTHT}](images/Schematic.png)
 
 `HomoPy` was developed to contain two commonly used homogenization methods, that is, the Mori-Tanka (cf. @Mori1973) for 3D stiffness predictions and a shear-lag modified Halpin-Tsai method (cf. @Cox1952 and @Halpin1969) for planar predictions, i.e. laminate predictions, which is based on the formulations in @Fu2019. The goal of `HomoPy` is to provide an open-source implementation of these methods with a particular focus on hybrid FRP modeling. Other modules offering similar functionalities are available, e.g. `fiberpy` and `mechmean` (cf. @mechmean), but to the author's knowledge none of them provide the capability to model hybrid FRPs consisting of different fiber materials and/or geometries.\
-A key element in HomoPy is the implementation of the graphical representation of the effective directional stiffnesses according to @Boehlke2001. Comparing different material systems or FRP tape layup orientations is intuitively easier with a graphical representation than comparing up to 21 stiffness components each. The package `mechkit` (cf. @Bauer2022b) also offers this visualization procedure, but is further capable of visualizing the generalized bulk modulus. Other visualization tools can be found in the package `continuum-mechanics` (cf. @continuum_mechanics), though the visualization is limited to second rank tensors. To the authors' knowledge, only `HomoPy` offers a visualization tool for elastic tensors in laminates and 2D stiffness tensors in general, respectively.
+A key element in HomoPy is the implementation of the graphical representation of the effective directional stiffnesses according to @Boehlke2001. Comparing different material systems or FRP tape layup orientations is intuitively easier with a graphical representation than comparing up to 21 stiffness components each. The package `mechkit` (cf. @Bauer2022b) also offers this visualization procedure, but is further capable of visualizing the generalized bulk modulus. `fiberoripy` (cf. @fiberoripy) uses a different general algorithm to represent second and forth order tensors, focusing on fiber orientation tensors rather than stiffness tensors. Other visualization tools can be found in the package `continuum-mechanics` (cf. @continuum_mechanics), though the visualization is limited to second rank tensors. To the authors' knowledge, only `HomoPy` offers a visualization tool for elastic tensors in laminates and 2D stiffness tensors in general, respectively.
 
 To this point, `HomoPy` is limited to calculating the effective elastic properties with the two methods mentioned above. Possible extensions for the future include thermal expansion properties and other homogenization methods.
 
 # Acknowledgements
 
 The research documented in this manuscript has been funded by the German Research Foundation (DFG) within the International Research Training Group “Integrated engineering of continuous-discontinuous long fiber-reinforced polymer structures” (GRK 2078/2). The support by the German Research Foundation (DFG) is gratefully acknowledged.
+
+In addition, N.C. acknowledges Julian Karl Bauer for valuable technical discussions.
 
 # References
