### 2023-04-25 67e2dd4
3D plot now has color mapping as a function of distance to origin
