### 2023-05-01 cbc8614
Add 'Scope and limitations' section in docs
