+++ docs/source/references.rst
@@ -1,20 +1,25 @@
 References
 ==========
 
-.. [Fu1998] Fu, S.Y. and Lauke, B. (1998), 'The elastic modulus of misaligned short-fiber-reinforced polymers', *Composites Science and Technology*, 58 (3), pp. 389-400, Available at: https://doi.org/10.1016/S0266-3538(97)00129-2
-.. [Christensen2012] Christensen, R. M. (2012), *Mechanics of Composite Materials*, Dover Publications
-.. [Chawla2019] Chawla, K. K. (2019), *Composite Materials*, Springer International Publishing
-.. [Mori1973] Mori, T. and Tanaka, K. (1973), 'Average stress in matrix and average elastic energy of materials with misfitting inclusions', *Acta Metallurgica*, 21 (5), pp. 571-574, Available at: https://doi.org/10.1016/0001-6160(73)90064-3
-.. [Cox1952] Cox, H. L. (1952), 'The elasticity and strength of paper and other fibrous materials', *British Journal of Applied Physics*, 3 (3), pp. 72–79, Available at: https://doi.org/10.1088/0508-3443/3/3/302
-.. [Halpin1969] Halpin, John C. (1969), 'Effects of environmental factors on composite materials', *Air Force Materials Laboratory, Air Force Systems Command*
+.. [Advani1987] Advani, S.G. and Tucker III, C.L. (1987), 'The Use of Tensors to Describe and Predict Fiber Orientation in Short Fiber Composites', *Journal of Rheology*, pp. 751-784, Available at: https://doi.org/10.1122/1.549945
+.. [Benveniste1987] Benveniste, Y. (1987), 'A new approach to the application of Mori-Tanaka's theory in composite materials', *Mechanics of Materials*, pp. 147-157, Available at: https://doi.org/10.1016/0167-6636(87)90005-6 
 .. [Boehlke2001] Böhlke, T. and Brüggemann, C. (2001), 'Graphical representation of the generalized Hooke's law', *Technische Mechanik*, 21 (2), pp. 145-158
 .. [Brannon2018] Brannon, R. M. (2018), 'Voigt and Mandel components', in *Rotation, Reflection, and Frame Changes*, IOP Publishing
-.. [Benveniste1987] Benveniste, Y. (1987), 'A new approach to the application of Mori-Tanaka's theory in composite materials', *Mechanics of Materials*, pp. 147-157, Available at: https://doi.org/10.1016/0167-6636(87)90005-6 
 .. [Brylka2017] Brylka, B. (2017), *Charakterisierung und Modellierung der Steifigkeit von langfaserverstärktem Polypropylen*, KIT Scientific Publishing
+.. [Chawla2019] Chawla, K. K. (2019), *Composite Materials*, Springer International Publishing
+.. [Christensen2012] Christensen, R. M. (2012), *Mechanics of Composite Materials*, Dover Publications
+.. [Cox1952] Cox, H. L. (1952), 'The elasticity and strength of paper and other fibrous materials', *British Journal of Applied Physics*, 3 (3), pp. 72–79, Available at: https://doi.org/10.1088/0508-3443/3/3/302
+.. [Eshelby1957] Eshelby, J.D. (1957), 'The determination of the elastic field of an ellipsoidal inclusion, and related problems', *Proceedings of the Royal Society of London*, A 241 (1957), pp. 376–396, Available at: https://doi.org/10.1098/rspa.1957.0133
[…]
