+++ README.md
@@ -26,7 +26,7 @@ Currently, HomoPy offers two types of homogenization procedures:
 **Figure 1:** Schematic of implemented homogenization methods. Inspired by [[Fu1998]](#Fu1998).
 
 ## Halpin-Tsai
-The Halpin-Tsai method is an empirical approach to homogenize two isotropic materials (cf. [[Halpin1969]](#Halpin1969)). Our approach is modified with the Shear-Lag model after Cox (cf. [[Cox1952]](#Cox1952)), which is also used in [[Fu2002]](#Fu2002) and [[Lauke2019]](#Lauke2019). Being a scalar homogenization scheme, it allows to calculate the effective stiffness in the plane which is orthogonal to the isotropic plane within transverse isotropic materials, as it is the case for unidirectional reinforced polymers. The effective stiffness, or Young's modulus, is then a function of the angle to the reinforcing direction. A fiber distrubtion within the plane is recognized by volume averaging of imaginary plies of individual orientations in analogy to the laminate theory.
+The Halpin-Tsai method is an empirical approach to homogenize two isotropic materials (cf. [[Halpin1969]](#Halpin1969)). Our approach is modified with the Shear-Lag model after Cox (cf. [[Cox1952]](#Cox1952)), which is also used in [[Fu2002]](#Fu2002) and [[Fu2019]](#Fu2019). Being a scalar homogenization scheme, it allows to calculate the effective stiffness in the plane which is orthogonal to the isotropic plane within transverse isotropic materials, as it is the case for unidirectional reinforced polymers. The effective stiffness, or Young's modulus, is then a function of the angle to the reinforcing direction. A fiber distrubtion within the plane is recognized by volume averaging of imaginary plies of individual orientations in analogy to the laminate theory.
 
 ## Mori-Tanaka
 The Mori-Tanaka scheme goes back to Mori and Tanaka (cf. [[Mori1973]](#Mori1973)) and is a mean-field homogenization scheme based on Eshelby's solution (cf. [[Eshelby1957]](#Eshelby1957)). The implementation so far allows UD (needle), spheroidal and circular inclusions. Our algorithm allows to homogenize materials with different types of fibers/inclusions, each possibily having an individual orientation distrubtion. Being a tensorial homogenization scheme, the fiber orientation tensor is directly included in the calculation and the result is an effective stiffness tensor. The authors would like to emphasize that the classic formulation after [[Benveniste1987]](#Benveniste1987) results in an effective stiffness tensor which violates thermodynamic requirements, i.e. which does not contain the major symmetry, for when multi-inclusion materials or non-isotropic inclusions are used, respectively. Further readings on this attribute are given in [[Qiu1990]](#Qiu1990) and [[Weng1990]](#Weng1990). To compensate this, HomoPy offers an algorithm introduced in [[Segura2023]](#Segura2023), which always results in symmetric effective stiffnesses.
@@ -60,7 +60,7 @@ Further topic related methods:
 <a id="Cox1952">[Cox1952]</a> H. L. Cox, *The elasticity and strength of paper and other fibrous materials*, British Journal of Applied Physics 3 (3) (1952) 72–79. doi:10.1088/05083443/3/3/302. \
 <a id="Fu1998">[Fu1998]</a> S.-Y. Fu, B. Lauke, *The elastic modulus of misaligned short-fiber-reinforced polymers*, Composites Science and Technology 58 (3) (1998) 389-40. doi:10.1016/S0266-3538(97)00129-2. \
[…]
