### 2023-05-01 #5295 @Extraweich
> > @lizarett thank you for your comment. Could you clarify a bit more to what extend you expect a derivation in the docs? The Git repository has a section for each homogenization procedure used in HomoPy in the README. Within these sections, the scope is mentioned and papers are referenced for the interested user to gain further insight into the reasons for limitations. My suggestion would be to add these sections to the docs as well and explicitly state the limitations (shapes, material symmetry, number of inclusions, etc.).
>
> Yes, this sounds good. Thank you.

I added a "Scope and limitations" section to the docs. I would be happy if you could check it out and see if this suits your suggestions :)
