### 2023-05-24 #5295 @Extraweich
> @lizarett thanks. Let me know if you are happy to formally recommend acceptance. If you do think that example should run I would encourage you to post an issue on that and get that resolved first. Let me know which route you feel is best. Thanks

The example requires to install another Python package, which is not required in HomoPy directly.
If it helps, I could reformulate the example with a fixed orientation tensor of 4th order to avoid the installation.
