+++ homopy/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.11"
+__version__ = "1.0.12"
 __author__ = "Nicolas Christ"
