### 2023-06-30 #5295 @nicoguaro
@Kevin-Mattheus-Moerman, I have marked all the boxes in the checklist. I think that the software is useful and we can move forward with the publishing stage.

I have some minor comments:

- There are 3 notebooks with examples in the repository. The one regarding Mori-Tanaka methods is quite complete while the other could be enriched with more text to provide explanations for people not familiar with the software/techniques.
- Continuum Mechanics does visualize fourth-rank tensors. It visualizes wave (phase) speeds or Christoffel tensors (see [the docs](https://continuum-mechanics.readthedocs.io/en/latest/usage.html#elasticity-tensor-visualization)). This is a different approach to the one used by the author in the software, though.
