### 2023-06-30 #5295 @Extraweich
@nicoguaro, thank you for finishing the review. I will address both your comments (hopefully tomorrow) soon. I plan on adding an explanatory notebook for Halpin-Tsai as well as change the lines in the paper concerning Continuum Mechanics to a correct statement.
