### 2023-06-30 55bcb00
Added logo to paper
