+++ paper/paper.md
@@ -39,6 +39,8 @@ The Python package `HomoPy` is a numerical software tool which provides computat
 The key element of `HomoPy` is the calculation and visualization of effective elastic stiffness properties of hybrid materials, i.e. multi-inclusion composites, using homogenization procedures. The current homogenization implementations are the conventional, three-dimensional Mori-Tanaka approach (cf. @Mori1973) in the formulation of @Benveniste1987 with a possible orientation averaging scheme after @Advani1987. A comprehensive study on the effects of the orientation averaging on homogenization procedures can be found in @Bauer2022a. To circumvent effective stiffness tensors, which are not major-symmetric and therefore violate thermodynamical principles, the algorithm in @Segura2023 was implemented and can be activated by a flag parameter to ensure symmetric stiffnesses. Alternatively, the shear-lag modified Halpin-Tsai approach (cf. @Fu2019) for purely planar information, for which the laminate theory is used to calculate effective stiffness properties, is available. The use field of these tools is academic research. An illustration of the implemented procedures is given in \autoref{fig:MTHT}. \
 Furthermore, `HomoPy` has the functionality to visualize the effective stiffness properties following @Boehlke2001, which allows for an intuitive comparison of different composites. The resulting effective properties can then be used in consequent numerical simulations.
 
+![Logo of HomoPy. \label{fig:HP}](images/HomoPy.png)
+
 # Statement of need
 
 Current ecological developments call for technical solutions to reduce the CO2 emissions of future innovations. Significant potential for promoting a better eco-balance lies in the development of new material systems, especially in the field of lightweight construction. Fiber-reinforced polymers (FRP) are a promising class of materials in the field of lightweight construction. The fibers used have high specific strength and stiffness properties, so that less material is required to achieve comparable properties compared with conventional materials, e.g. steel. The polymer serves to hold the fibers in place, transfer stresses between them, and increase the toughness properties of the composite. For more literature on FRP, see @Christensen1980 and @Chawla2019.
