### 2023-06-30 5d5354b
Changed logo size in paper
