### 2023-07-01 #5295 @likask
@Extraweich @Kevin-Mattheus-Moerman

Hello, your code is really helpful and I plan to use it. You're using tried-and-true methods backed up by a lot of research. That said, I think it would be even better if, in the future, you could add examples that use experimental data to notebooks, and compare overall estimates.

I also want to suggest another improvement. Could you add a feature for stochastic analysis? Maybe an example showing how to take into account the uncertainties about the shape of the composite and the properties of its constituents.

I'm sorry it took me a while to finish this review.
