### 2023-07-01 #5295 @Extraweich
@likask, thank you for your feedback and review. The data in the notebooks (usually carbon fiber reinforced polyamide 6) stems from my own research in the field. Currently we are working on a publication which uses HomoPy to compare its results to experimental investigations. Once this is published, I will add a reference to the project.
Concerning the stochastic analysis, I agree that this would be an interesting feature and would improve the functionality of HomoPy. Technically, the input could already be realized in a stochastic sense, if you correlate the probability of a given characteristic with its volume fraction in the composite. The output would be purely deterministic, though, leaving no interpretation of standard deviations. I think working on this could be a good extension to HomoPy in the future, but I would need to read more into the field of stochastic homogenization methods.

I would like to thank all of you for the time you invested in reviewing this project. :)
