### 2023-07-01 #5295 @Extraweich
@Kevin-Mattheus-Moerman, before the publishing stage begins, would it be possible to change the referenced version from 1.0.10 to 1.1.0? Once you give me green lights, I would update the repository.
