+++ homopy/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.12"
+__version__ = "1.1.0"
 __author__ = "Nicolas Christ"
