+++ README.md
@@ -7,7 +7,7 @@
 [![Documentation status](https://readthedocs.org/projects/homopy/badge/?version=latest)](https://homopy.readthedocs.io/en/latest/?badge=latest)
 [![Codacy Badge](https://app.codacy.com/project/badge/Grade/ffc0c7b16d154bc18cccc1e857724d86)](https://www.codacy.com/gh/Extraweich/homopy/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=Extraweich/homopy&amp;utm_campaign=Badge_Grade)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Extraweich/homopy/main?labpath=%2Fexamples%2FHybrid.ipynb)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Extraweich/homopy/main?labpath=examples%2FInteractive_hybrid.ipynb)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7967631.svg)](https://doi.org/10.5281/zenodo.7967631)
 
 *Your solution for stiffness problems*
@@ -46,7 +46,7 @@ pip install homopy
 
 ### Interactive example
 
-An interactive example to intuitively see the effects of fiber distributions on the effective properties of hybrid materials can be found in [Binder](https://mybinder.org/v2/gh/Extraweich/homopy/main?labpath=%2Fexamples%2FHybrid.ipynb).
+An interactive example to intuitively see the effects of fiber distributions on the effective properties of hybrid materials can be found in [Binder](https://mybinder.org/v2/gh/Extraweich/homopy/main?labpath=examples%2FInteractive_hybrid.ipynb).
 
 <p align="center">
   <picture>
