### 2023-07-07 #5295 @Kevin-Mattheus-Moerman
@Extraweich side note, I created a PR which is not a requirement for this review but aims to fix a rendering issue with the logo in the README: https://github.com/Extraweich/homopy/pull/24
