+++ README.md
@@ -9,7 +9,7 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)
 [![status](https://joss.theoj.org/papers/3ff6841e9a89fc91a0578df9ca677179/status.svg)](https://joss.theoj.org/papers/3ff6841e9a89fc91a0578df9ca677179)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Extraweich/homopy/main?labpath=examples%2FInteractive_hybrid.ipynb)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7967631.svg)](https://doi.org/10.5281/zenodo.7967631)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8124344.svg)](https://doi.org/10.5281/zenodo.8124344)
 
 *Your solution for stiffness problems*
 
