### 2023-07-07 #5295 @Extraweich
@Kevin-Mattheus-Moerman, I checked all the points on the list. I guess that you need to cross them for me, though.
I added a new release, updated the Zenodo DOI and checked authors and the license.
